requirements.txt
Flask>=2.1.0
firebase-admin==6.5.0
google-cloud-storage==2.16.0
spacy==3.7.5
nltk==3.8.1
openai==1.38.0
pandas==2.2.1
gunicorn==22.0.0
requests==2.32.2
numpy==1.26.4
pdfplumber==0.11.4
tiktoken==0.7.0
langchain==0.2.9
langchain-core==0.2.23
langchain-community==0.2.10
langchain_openai
